feat: integrate quick and refined searches
